requirements.txt
psycopg2-binary
faker
faker-vehicle